requirements.txt
requests==2.31.0
pyTelegramBotAPI
python-dotenv
psycopg2-binary
feedparser
schedule
